feature/AIQA-1593 [GEO][DJANGO] widoki/kontrolery: każda klasa w osobnym pliku
