+++ backend-python/api/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
-from .models import River, City, Mountain
+from .models.river import River
+from .models.city import City
+from .models.mountain import Mountain
 
 admin.site.register(Mountain)
 admin.site.register(River)
