+++ backend-python/api/serializers/mountain_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from ..models import Mountain
+from ..models.mountain import Mountain
 
 
 class MountainSerializer(serializers.ModelSerializer):
