+++ backend-python/api/views/city/city_views.py
@@ -1,0 +1,9 @@
+from rest_framework import generics
+
+from ...models.city import City
+from ...serializers.city_serializer import CitySerializer
+
+
+class CityDetail(generics.RetrieveUpdateDestroyAPIView):
+    queryset = City.objects.all()
+    serializer_class = CitySerializer
